Switch FVcounter/PPcounter to INT(KIND=8)

+++ src/analyze/analyze_vars.f90
@@ -29,10 +29,10 @@ REAL                 :: WriteData_dt                      !< time intervall at w
 REAL                 :: tWriteData                        !< actual time at which next solution IO will be performed
 ! precomputed variables
 #if FV_ENABLED
-INTEGER              :: totalFV_nElems=0
+INTEGER(KIND=8)      :: totalFV_nElems=0
 #endif
 #if PP_LIMITER
-INTEGER              :: totalPP_nElems=0
+INTEGER(KIND=8)      :: totalPP_nElems=0
 #endif
 REAL,ALLOCATABLE     :: wGPSurf(:,:)                      !< wGPSurf(i,j)=wGP(i)*wGP(j)
 REAL,ALLOCATABLE     :: wGPVol(:,:,:)                     !< wGPVol(i,j,k)=wGP(i)*wGP(j)*wGP(k)
